WIP - moved download code block to profile summary component
