+++ src/app/+data-and-analytics/data-and-analytics.component.ts
@@ -302,87 +302,4 @@ export class DataAndAnalyticsComponent implements OnInit {
       .pipe(map(collection => (collection._embedded.individual as Individual[]).sort((a, b) => a.name.localeCompare(b.name))));
   }
 
-  public onSelectAll(event: Event, organization: any): void {
-    const checked = (event.target as HTMLInputElement).checked;
-    const people = organization.people || [];
-    if (checked) {
-      this.selectedPeopleSubject.next([...people]);
-    } else {
-      this.selectedPeopleSubject.next([]);
-    }
-    const checkboxes = document.querySelectorAll<HTMLInputElement>('.selected-profile-checkbox');
-    checkboxes.forEach(cb => cb.checked = checked);
-  }
-
-  public onSelectPerson(event: Event, person: Individual): void {
-    const checked = (event.target as HTMLInputElement).checked;
-    if (checked) {
-      const current = this.selectedPeopleSubject.value;
-      if (!current.find(p => p.id === person.id)) {
-        this.selectedPeopleSubject.next([...current, person]);
-      }
-    } else {
-      const current = this.selectedPeopleSubject.value.filter(p => p.id !== person.id);
-      this.selectedPeopleSubject.next(current);
-    }
-  }
-
-  public downloadSelectedPeople(organization: any): void {
-    this.route.queryParams.pipe(take(1)).subscribe((params) => {
-      const orgId = params?.selectedOrganization ? params.selectedOrganization : organization.id;
-      const exportName = params?.export.toLowerCase().replace(/ /g, '_');
-      console.log("\n exportName: ", exportName);
-      const selectedIds = this.selectedPeopleSubject.value.map(p => p.id);
-
-      if (!orgId) {
-        console.error('Download failure: Missing Organization id.');
-        return;
-      }
-
-      if (!selectedIds.length || selectedIds.length === (organization.people?.length ?? 0)) {
-        this.restService.get<Blob>(
-          organization._links[exportName].href,
-          { observe: 'response', responseType: 'blob' as 'json' })
-          .pipe(take(1))
-          .subscribe((response: any) => {
-            const contentDisposition = response.headers.get('Content-Disposition');
-            const filename = !!contentDisposition
-                             ? contentDisposition.match(/^.*filename=(.*)$/)[1] : 'export.zip';
-            this.download(response, filename);
-          },);
-      } else {
-          const updatedHref = `${this.appConfig.serviceUrl}/individual/${orgId}/export?type=zip&name=${encodeURIComponent(exportName)}`;
-          this.restService.post(
-            updatedHref,
-            selectedIds,
-            { observe: 'response', responseType: 'blob' as 'json', headers: { 'Content-Type': 'application/json' } }
-          ).subscribe({
-            next: (response: any) => {
-              const contentDisposition = response.headers.get('Content-Disposition');
-
-              const filename = !!contentDisposition
-                               ? contentDisposition.match(/^.*filename=(.*)$/)[1] : 'selected_profile(s).zip';
-              this.download(response, filename);
-            },
-            error: (err) => console.error('Failed to download selected profiles.', err),
-          });
-      }
-    });
-  }
-
-  private download(response: any, defaultFileName = 'profile_summary_download.zip') {
-    const blob = response.body || response;
-    const contentDisposition = response.headers?.get?.('Content-Disposition');
-    const filename = contentDisposition
-      ? contentDisposition.match(/^.*filename=(.*)$/)[1]
-      : defaultFileName;
-
-    const url = window.URL.createObjectURL(blob);
-    const anchor = document.createElement('a');
-    anchor.download = filename;
-    anchor.href = url;
-    anchor.click();
-    window.URL.revokeObjectURL(url);
-  }
-
 }
